seprated application logic and dom manipulation
