+++ index.html
@@ -9,21 +9,12 @@
   </head>
   <body>
      <h1 class="text-center main-logo">My Library</h1>
-    <!-- <label for="title">Title</label>
-    <input type="text" name="title" id="title" />
-    <label for="author">author</label>
-    <input type="text" name="author" id="author" />
-    <label for="pages">pages</label>
-    <input type="text" name="pages" id="pages" />
-    <label for="read">read</label>
-    <input type="text" name="read" id="read" />
-    <button type="submit" id= 'create'>Create Book</button> -->
     <button id="addBookBtn" class="center-it">Add Book</button>
-    <div id="form">
-
-    </div>
+    <form id="form">
+    
+    </form>
     
     <div class="work"></div>
-    <script src="./library.js"></script>
+    <script type="module" src="./js/dom.js"></script>
   </body>
 </html>
